Thread  interruption Java does not support safely forcing stop a thread. JAVA provides a cooperative mechanism that one thread ask another to stop what it is doing Thread interrupt in java is only a boolean flag to signal that a interrupt request is raised.  it is up to the thread/task/service itself to determine how to respond a interrupt request.  A interruption may have multiple recipients so we should not swallow the interruption that you may not know how many recipients it has
Thread  interruption
Java does not support safely forcing stop a thread.
JAVA provides a cooperative mechanism that one thread ask another to stop what it is doing
Thread interrupt in java is only a boolean flag to signal that a interrupt request is raised.  it is up to the thread/task/service itself to determine how to respond a interrupt request.

A interruption may have multiple recipients so we should not swallow the interruption that you may not know how many recipients it has